% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_articulation_point.R
\name{add_articulation_point}
\alias{add_articulation_point}
\title{Adds articulation point to an igraph graph
Returns a graph}
\usage{
add_articulation_point(graph)
}
\description{
Adds articulation point to an igraph graph
Returns a graph
}
