% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_network_descriptives.R
\name{add_network_descriptives}
\alias{add_network_descriptives}
\title{Add network descriptives to an igraph graph
Returns string of descriptives}
\usage{
add_network_descriptives(
  graph,
  order = "Order=",
  size = "Size=",
  clustering = "Clustering Coeffiecient=",
  diameter = "Diameter=",
  mean_distance = "Mean distance="
)
}
\description{
Add network descriptives to an igraph graph
Returns string of descriptives
}
