% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_distribution_summary.R
\name{degree_distribution_summary}
\alias{degree_distribution_summary}
\title{Helper function to get the degree distribution of the graph}
\usage{
degree_distribution_summary(graph)
}
\arguments{
\item{graph}{\itemize{
\item an igraph graph
}}
}
\value{
a tibble
}
\description{
Helper function to get the degree distribution of the graph
}
