% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tau.R
\name{estimate_tau}
\alias{estimate_tau}
\title{Estimates Kendall's tau}
\usage{
estimate_tau(x, y = NULL)
}
\arguments{
\item{x}{numeric vector or list}

\item{y}{numeric vector}
}
\value{
tau
}
\description{
Estimates Kendall's tau
}
