% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_inf_plot.R
\name{expected_inf_plot}
\alias{expected_inf_plot}
\title{Create Plot of Graph with expected influence metric}
\usage{
expected_inf_plot(network, toolname)
}
\arguments{
\item{network}{igraph graph}

\item{toolname}{string name of node,}
}
\value{
plot of graph enriched with expected influnce
}
\description{
Create Plot of Graph with expected influence metric
}
