% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subscale.R
\name{find_subscale}
\alias{find_subscale}
\title{Article-specific helper function to match original category of a subscale item to predicted}
\usage{
find_subscale(original, predicted)
}
\arguments{
\item{original}{string vector of true membership names}

\item{predicted}{string vector of predicted membership of names}
}
\value{
list of output stats
}
\description{
Article-specific helper function to match original category of a subscale item to predicted
}
