% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridge_estimate.R
\name{get_bridge_estimate}
\alias{get_bridge_estimate}
\title{Gets bridge influence statistics}
\usage{
get_bridge_estimate(
  network,
  dec = 2,
  seed = 1234,
  method = "fast_greedy",
  weights = TRUE
)
}
\arguments{
\item{network}{igraph graph}

\item{dec}{\itemize{
\item integer - decimal count
}}

\item{seed}{random seed}

\item{weights}{boolen - if TRUE extracts the weight of the network and uses it in community detection}

\item{methid}{string name of method (spinglass, fast_greedy, walktrap, optimal)}
}
\value{
character vector
}
\description{
Gets bridge influence statistics
}
