% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_names.R
\name{get_item_names}
\alias{get_item_names}
\title{Gets row names}
\usage{
get_item_names(covMatrix, labels, to_replace = c("PCA."))
}
\arguments{
\item{covMatrix}{data frame}

\item{labels}{vector of string names}

\item{to_replace}{optional string vector to be replaced}
}
\value{
character vector
}
\description{
Gets row names
}
