% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_itemnames_to_tibble.R
\name{get_itemnames}
\alias{get_itemnames}
\title{Transforms name vector to uniquie tibble}
\usage{
get_itemnames(namevector)
}
\arguments{
\item{namevector}{string vector of names}
}
\description{
Transforms name vector to uniquie tibble
}
