% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_descriptives.R
\name{get_network_descriptives}
\alias{get_network_descriptives}
\title{Gets network descriptives to an igraph graph
Returns string of descriptives}
\usage{
get_network_descriptives(
  graph,
  order = "Order=",
  size = "Size=",
  clustering = "Clustering Coeffiecient=",
  diameter = "Diameter=",
  mean_distance = "Mean distance="
)
}
\description{
Gets network descriptives to an igraph graph
Returns string of descriptives
}
