% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TPR.R
\name{get_TPR}
\alias{get_TPR}
\title{Computes the True Positive Rates from the result of simulate_communities()}
\usage{
get_TPR(simulatedCommunities, graph)
}
\arguments{
\item{simulatedCommunities}{\itemize{
\item result of simulate_communities()
}}

\item{graph}{\itemize{
\item an igraph graph
}}
}
\value{
a tibble of TPR statistics
}
\description{
Computes the True Positive Rates from the result of simulate_communities()
}
