% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bridge_plot.R
\name{make_bridge_plot}
\alias{make_bridge_plot}
\title{Create Plot of Graph with bridge influence metric}
\usage{
make_bridge_plot(network, toolname)
}
\arguments{
\item{network}{igraph graph}

\item{toolname}{string name of node,}
}
\value{
plot of graph enriched with bridge influnce
}
\description{
Create Plot of Graph with bridge influence metric
}
