% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_community_graph.R
\name{make_community_graph}
\alias{make_community_graph}
\title{Makes qgraph graph with communities}
\usage{
make_community_graph(graph, covMatrix, labels)
}
\arguments{
\item{graph}{An igraph graph}

\item{covMatrix}{Covariance matrix of the underlying parameters}

\item{labels}{string vector of node labels}
}
\value{
qraph graph
}
\description{
Makes qgraph graph with communities
}
