% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_accuracy_plot.R
\name{network_accuracy_plot}
\alias{network_accuracy_plot}
\title{Helper function to make a network accuracy plot}
\usage{
network_accuracy_plot(accuracy, tool = "", multiple = FALSE, save_data = FALSE)
}
\description{
Helper function to make a network accuracy plot
}
