% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stability_estimator.R
\name{network_stability_estimator}
\alias{network_stability_estimator}
\title{Estimates stability statistics of a network}
\usage{
network_stability_estimator(network)
}
\arguments{
\item{network}{igraph graph}
}
\value{
stability comptation results
}
\description{
Estimates stability statistics of a network
}
