% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stability_plot.R
\name{network_stability_plot}
\alias{network_stability_plot}
\title{Visualizes the result of network stability estimation}
\usage{
network_stability_plot(stability_estimate)
}
\arguments{
\item{stability_estimate}{\itemize{
\item result of network_stability_estimator()
}}
}
\value{
Plot
}
\description{
Visualizes the result of network stability estimation
}
