% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_communities.R
\name{simulate_communities}
\alias{simulate_communities}
\title{Function to compute various communities by various community detection methods}
\usage{
simulate_communities(graph, i = 1000, path = "simulate.RDS")
}
\arguments{
\item{graph}{\itemize{
\item an igraph graph
}}

\item{i}{\itemize{
\item integer - number of runs
}}

\item{path}{string - destination file to write results}
}
\value{
a tibble
}
\description{
Function to compute various communities by various community detection methods
}
